app/src/main/java/com/doug/cashflow/view/outputs/Savings.java
package com.doug.cashflow.view.outputs;

import android.content.Context;

import com.doug.cashflow.R;
import com.doug.cashflow.model.system.ResultsDataNode;

import java.util.ArrayList;

public class Savings extends Account implements Accounts {
    private static final Integer title = R.string.output_header_savings;

    private Context mContext;
    private final Utils    utils = new Utils();

    public Savings(Context context, ArrayList<ResultsDataNode> values) {
        super(context, values, title);

        mContext = context;
    }

    public String getName() {
        return utils.getString(mContext, title);
    }
}
